pyproject.toml
[project]
name = "yantrikdb"
version = "0.1.0"
description = "A Cognitive Memory Engine for Persistent AI Systems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pranab Sarkar", email = "developer@pranab.co.in" },
]
dependencies = [
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "uuid-utils>=0.9.0",
    "mcp[cli]>=1.2.0",
    "click>=8.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-benchmark>=4.0.0",
]
sync = [
    "websockets>=12.0",
]
langchain = [
    "langchain-core>=0.2.0",
]
openai = [
    "openai>=1.0.0",
]
crewai = [
    "crewai>=0.40.0",
]
agent = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
yantrikdb = "yantrikdb.cli:main"
yantrikdb-mcp = "yantrikdb.mcp:main"
yantrikdb-server = "yantrikdb.api:main"
yantrikdb-sync = "yantrikdb.sync.transport:main"
yantrik-companion = "yantrikdb.agent.service:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"yantrikdb.agent" = ["static/**/*"]
